refactor factory to separate file
